Convert `entrypoints/sign_up` to Typescript
